🔒️ Added a secure mutation wrapper
We discussed that this shouldn't recieve attention yet, but it was much
faster to write the logic once here than copy-paste all over the
codebase.

If it doesn't work as expected then I'll strip it out - as we have other
priorities. I don't see why it shouldn't work though, types are
identical and logic is taken straight from the area form.